requirements.txt
asgiref==3.5.2
beautifulsoup4==4.11.1
black==22.3.0
bootstrap4==0.1.0
certifi==2022.6.15
charset-normalizer==2.0.12
click==8.1.3
Django==4.0.5
django-autocomplete-light==3.9.4
django-bootstrap-datepicker-plus==4.0.0
django-bootstrap-v5==1.0.11
django-bootstrap4==22.1
django-js-asset==2.0.0
django-modeltranslation==0.18.2
django-mptt==0.13.4
django-rosetta==0.9.8
django-taggit==3.0.0
djangorestframework==3.13.1
idna==3.3
mypy-extensions==0.4.3
pathspec==0.9.0
Pillow==9.1.1
platformdirs==2.5.2
polib==1.1.1
pytz==2022.1
requests==2.28.0
six==1.16.0
soupsieve==2.3.2.post1
sqlparse==0.4.2
tomli==2.0.1
urllib3==1.26.9
